backend/model_service/hf_resnet50/requirements.txt
Flask >= 2.2
mysql-connector-python
Pillow
numpy
flask-cors
prometheus-client
python-dotenv